requirements.txt
accelerate==0.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bigdl-core-xe-all==2.7.0b20250616
bitsandbytes==0.43.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==2.19.0
dill==0.3.8
docstring_parser==0.16
dpcpp-cpp-rt==2025.0.4
filelock==3.18.0
fire==0.7.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.3.1
huggingface-hub==0.33.0
idna==3.10
intel-cmplr-lib-rt==2025.0.4
intel-cmplr-lib-ur==2025.0.4
intel-cmplr-lic-rt==2025.0.4
intel-opencl-rt==2025.0.4
intel-openmp==2025.0.4
intel-sycl-rt==2025.0.4
intel_extension_for_deepspeed @ git+https://github.com/intel/intel-extension-for-deepspeed.git@ec332772d8bb71c6d89804a862f039fc91ae96b5
intel_extension_for_pytorch==2.6.10+xpu
ipex-llm==2.3.0b20250616
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mkl==2025.0.1
mkl-dpcpp==2025.0.1
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
networkx==3.5
numpy==1.26.4
onednn==2025.0.1
onednn-devel==2025.0.1
onemkl-sycl-blas==2025.0.1
onemkl-sycl-datafitting==2025.0.1
onemkl-sycl-dft==2025.0.1
onemkl-sycl-lapack==2025.0.1
onemkl-sycl-rng==2025.0.1
onemkl-sycl-sparse==2025.0.1
onemkl-sycl-stats==2025.0.1
onemkl-sycl-vm==2025.0.1
packaging==25.0
pandas==2.3.0
peft==0.10.0
pillow==11.2.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.6.0rc0
scipy==1.13.0
sentencepiece==0.2.0
shtab==1.7.2
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tbb==2022.1.0
tcmlib==1.2.0
termcolor==3.1.0
tokenizers==0.15.2
torch==2.6.0+xpu
torchaudio==2.6.0+xpu
torchvision==0.21.0+xpu
tqdm==4.67.1
transformers==4.36.1
trl==0.8.6
typeguard==4.4.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.9.24
tzdata==2025.2
umf==0.9.1
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.1